+++ diskord-bot/src/commonMain/kotlin/com/jessecorbett/diskord/bot/interaction/ResponseContext.kt
@@ -5,19 +5,28 @@ import com.jessecorbett.diskord.api.channel.Embed
 import com.jessecorbett.diskord.api.common.ActionRow
 import com.jessecorbett.diskord.api.common.Attachment
 import com.jessecorbett.diskord.api.common.TextInput
+import com.jessecorbett.diskord.api.interaction.CommandInteractionDataResolved
+import com.jessecorbett.diskord.api.interaction.CommandInteractionOptionResponse
 import com.jessecorbett.diskord.api.interaction.Interaction
 import com.jessecorbett.diskord.api.interaction.ModalSubmit
-import com.jessecorbett.diskord.api.interaction.callback.ChannelMessageWithSource
-import com.jessecorbett.diskord.api.interaction.callback.CreateModalResult
-import com.jessecorbett.diskord.api.interaction.callback.DeferredChannelMessageWithSource
-import com.jessecorbett.diskord.api.interaction.callback.InteractionCommandCallbackDataFlags
-import com.jessecorbett.diskord.api.interaction.callback.InteractionResponse
+import com.jessecorbett.diskord.api.interaction.callback.*
 import com.jessecorbett.diskord.api.webhook.CreateWebhookMessage
 import com.jessecorbett.diskord.bot.BotContext
 
-public class ResponseContext internal constructor(
+/**
+ * The action scope of the response to an [Interaction].
+ *
+ * @property command The instance of the command the interaction is for
+ * @property options The parameters passed from the Discord user via the command
+ * @property data The resolved data relative to the options
+ */
+@InteractionModule
+public data class ResponseContext<I: Interaction> internal constructor(
     private val context: BotContext,
-    private val registerModal: (String, suspend ResponseContext.(ModalSubmit) -> Unit) -> Unit
+    public val command: I,
+    public val options: List<CommandInteractionOptionResponse>,
+    public val data: CommandInteractionDataResolved?,
+    private val registerModal: (String, suspend ResponseContext<ModalSubmit>.(ModalSubmit) -> Unit) -> Unit
 ) : BotContext by context {
     private var hasResponded = false
     private var hasAckedForFuture = false
@@ -25,7 +34,7 @@ public class ResponseContext internal constructor(
     /**
      * Acknowledges the interaction and informs the user and Discord that a response will be coming shortly
      */
-    public suspend fun Interaction.acknowledgeForFutureResponse(ephemeral: Boolean = false) {
+    public suspend fun acknowledgeForFutureResponse(ephemeral: Boolean = false) {
         hasAckedForFuture = true
         respond(DeferredChannelMessageWithSource(DeferredChannelMessageWithSource.Data(
             if (ephemeral) InteractionCommandCallbackDataFlags.EPHEMERAL else InteractionCommandCallbackDataFlags.NONE
@@ -35,7 +44,7 @@ public class ResponseContext internal constructor(
     /**
      * Responds to the interaction with a channel message
      */
-    public suspend fun Interaction.respond(builder: suspend ResponseBuilder.() -> Unit) {
+    public suspend fun respond(builder: suspend ResponseBuilder.() -> Unit) {
         val response = ResponseBuilder().apply {
             builder()
         }.build()
@@ -50,11 +59,14 @@ public class ResponseContext internal constructor(
         client.deleteOriginalInteractionResponse()
     }
 
+    /**
+     * Respond to the interaction by creating a modal for the user
+     */
     @InteractionModule
     public suspend fun Interaction.createModal(
         title: String,
         vararg inputs: TextInput,
-        callback: suspend ResponseContext.(ModalSubmit) -> Unit
+        callback: suspend ResponseContext<ModalSubmit>.(ModalSubmit) -> Unit
     ) {
         val response = CreateModalResult(CreateModalResult.Data(
             customId = id,
@@ -66,10 +78,10 @@ public class ResponseContext internal constructor(
         registerModal(id, callback)
     }
 
-    private suspend fun Interaction.respond(response: InteractionResponse) {
+    private suspend fun respond(response: InteractionResponse) {
         if (hasAckedForFuture && response is ChannelMessageWithSource) {
-            context.webhook(applicationId).execute(
-                token,
+            context.webhook(command.applicationId).execute(
+                command.token,
                 CreateWebhookMessage(
                     useTTS = response.data.tts,
                     content = response.data.content,
@@ -85,13 +97,18 @@ public class ResponseContext internal constructor(
 
 
         if (hasResponded) {
-            client.deleteOriginalInteractionResponse()
+            command.client.deleteOriginalInteractionResponse()
         }
 
         hasResponded = true
-        client.createInteractionResponse(id, response)
+        command.client.createInteractionResponse(command.id, response)
     }
 
+    /**
+     * Convenience class for building a message response
+     *
+     * TODO: Merge into one convenience class for all message building
+     */
     public class ResponseBuilder {
         public var content: String? = null
         public var tts: Boolean = false