+++ diskord-testbot-jvm/src/main/kotlin/com.jessecorbett.diskord.testbot/Bot.kt
@@ -45,30 +45,48 @@ suspend fun main() {
         }
 
         interactions {
-            userCommand("print") { it, data ->
-                it.respond {
-                    content = "Test data for interaction " + data.convertedUsersRolesChannels
+            userCommand("print") {
+                respond {
+                    content = "Test data for interaction $data"
                     ephemeral
                 }
             }
 
-            messageCommand("test") { interaction, data ->
-                println(interaction)
-            }
-
             slashCommand("echo", "Makes the bot say something") {
                 val message by stringParameter("message", "The message")
-                callback { interaction, _ ->
-                    interaction.respond {
+                callback {
+                    respond {
                         content = message
                     }
                 }
             }
 
             slashCommand("timestamp", "Prints the current timestamp") {
-                callback { interaction, _ ->
-                    interaction.respond {
+                callback {
+                    respond {
                         content = Clock.System.now().toString()
+                    }
+                }
+            }
+
+            commandGroup("test", "Test group", guildId = "424046347428167688") {
+                subgroup("foo", "Foo test group") {
+                    slashCommand("bar", "Bar command") {
+                        val msg by stringParameter("message", "The message to say")
+                        callback {
+                            respond {
+                                content = "$msg bar!"
+                            }
+                        }
+                    }
+                }
+
+                slashCommand("fizz", "Buzz command") {
+                    val msg by stringParameter("message", "The message to say")
+                    callback {
+                        respond {
+                            content = "$msg buzz!"
+                        }
                     }
                 }
             }